int->uint64_t so longer simulations are possible

+++ Quantom/Quantom/Analyzer.cuh
@@ -34,8 +34,8 @@ public:
 
 	AnalyzedPackage analyzeEnergy(Simulation* simulation); // Prints a file of doubles: [step, molecule, atom, coordinate_dim]
 
-	Float3* analyzeSolvateEnergy(Simulation* simulation, int n_steps);
-	Float3* analyzeCompoundEnergy(Simulation* simulation, int n_steps);
+	Float3* analyzeSolvateEnergy(Simulation* simulation, uint64_t n_steps);
+	Float3* analyzeCompoundEnergy(Simulation* simulation, uint64_t n_steps);
